feat: cleaner code
