+++ components/AccountDetails.vue
@@ -1,0 +1,9 @@
+<script setup lang="ts">
+</script>
+
+<template>
+
+</template>
+
+<style scoped lang="scss">
+</style>
