+++ pages/YourAccount.vue
@@ -17,53 +17,26 @@ const firestoreClient = ref({
     ? userProfile.value.photoURL
     : "raiden.png",
   email: userProfile.value.email,
-  phone: userProfile.value.providerData[0].phoneNumber
-    ? userProfile.value.providerData[0].phoneNumber
+  phone: userProfile.value.phoneNumber
+    ? userProfile.value.phoneNumber
     : "",
 });
 const useAuth = useUser();
 
-const modalMessage = ref();
-const firstTimePhone = ref<string>("");
-const showModal = ref<boolean>(false);
+
 const buttonCaptcha = ref<boolean>(false);
 const changePhone = ref<boolean>(false);
 const newNumber = ref<string>("");
 const verCode = ref<string>("");
 const newProfilePicture = ref(["keqing.png", "kokomi.png", "raiden.png"]);
-let verificationId: string;
-const applicationVerifier = useState<RecaptchaVerifier>("captcha");
+
 
 async function updatePhoto(photo: string) {
   firestoreClient.value.avatar = photo;
   await updatePicture(photo);
 }
 
-async function updatePhone(number: string) {
-  number = number.replace(/\s+/g, "");
 
-  //Button animates on Captcha loading
-  buttonCaptcha.value = true;
-  //Provider of auth
-  const provider = new PhoneAuthProvider(useAuth.value);
-  try {
-    //Starts verification process
-    const test = await applicationVerifier.value.verify();
-
-    if (test) {
-      verificationId = await provider.verifyPhoneNumber(
-        number,
-        applicationVerifier.value
-      );
-    }
-
-    //Shows modal for verification code
-    modalMessage.value = "Enter verification code";
-  } catch (err) {
-    modalMessage.value = "Too many requests!";
-  }
-  showModal.value = true;
-}
 
 async function verifyCode(verificationCode: any) {
   //Hides modal
@@ -164,69 +137,7 @@ async function deletePhone() {
   <NavBar :isAccount="true" :av="firestoreClient"></NavBar>
   <div class="flex-center">
     <h2>Account</h2>
-    <div class="account">
-      <div class="category"><h3>MEMBSERSHIP AND BILLING</h3></div>
-      <div class="details">
-        <div class="detail-item">
-          <p>{{ firestoreClient.email }}</p>
-          <button class="change-button">Change your email address</button>
-        </div>
-        <div class="detail-item">
-          <p>Password: ********</p>
-          <button class="change-button">Change password</button>
-        </div>
-
-        <div
-          class="detail-item"
-          :class="!firestoreClient.phone ? 'item-input' : ''"
-        >
-          <form
-            class="item-input"
-            ref="phoneForm"
-            @submit.prevent="updatePhone(firstTimePhone)"
-            v-if="!firestoreClient.phone"
-            action=""
-          >
-            <BaseInput
-              :border="true"
-              :background="true"
-              name="Phone Number"
-              type="tel"
-              pattern="^\+\d{1,4}(\s?\d){9,12}$"
-              v-model="firstTimePhone"
-              required
-            />
-
-            <button
-              id="recaptcha-container"
-              :class="buttonCaptcha ? 'button-spinner' : ''"
-              class="change-button"
-            >
-              <span v-if="!buttonCaptcha">Change phone number</span>
-              <div v-else class="loader">
-                <span class="loader-circle"></span>
-              </div>
-            </button>
-          </form>
-
-          <p v-else>{{ firestoreClient.phone }}</p>
-
-          <div class="btn" v-if="firestoreClient.phone">
-            <button
-              id="recaptcha-container"
-              @click="updatePhone(firestoreClient.phone)"
-              :class="buttonCaptcha ? 'button-spinner' : ''"
-              class="change-button"
-            >
-              <span v-if="!buttonCaptcha">Change phone number</span>
-              <div v-else class="loader">
-                <span class="loader-circle"></span>
-              </div>
-            </button>
-          </div>
-        </div>
-      </div>
-    </div>
+   
     <div class="delete">
       <button
         id="recaptcha-container"
