+++ server/api/getGenres.ts
@@ -1,4 +1,4 @@
-import { Movie, Movies } from "~/types";
+
 import type { H3Event } from "h3";
 export default defineEventHandler(async (event: H3Event) => {
   const config = useRuntimeConfig();
